+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-*No unreleased changes*
+### Added
+- Add a button to remove shortcuts in settings
+
+### Changed
+- Improved appearance of shortcuts settings
 
 ## [1.0.2] - 2018-07-04
 ### Added